Refactor: Separate execution indexing logic from ZestGuidance
