+++ fuzz/src/main/java/edu/berkeley/cs/jqf/fuzz/ei/ZestGuidance.java
@@ -48,7 +48,6 @@ import java.util.Deque;
 import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Iterator;
-import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Random;
@@ -56,18 +55,12 @@ import java.util.Set;
 import java.util.concurrent.TimeUnit;
 import java.util.function.Consumer;
 
-import edu.berkeley.cs.jqf.fuzz.ei.ExecutionIndex.Prefix;
-import edu.berkeley.cs.jqf.fuzz.ei.ExecutionIndex.Suffix;
 import edu.berkeley.cs.jqf.fuzz.guidance.Guidance;
 import edu.berkeley.cs.jqf.fuzz.guidance.GuidanceException;
 import edu.berkeley.cs.jqf.fuzz.guidance.Result;
 import edu.berkeley.cs.jqf.fuzz.guidance.TimeoutException;
 import edu.berkeley.cs.jqf.fuzz.util.Coverage;
-import edu.berkeley.cs.jqf.fuzz.util.ProducerHashMap;
-import edu.berkeley.cs.jqf.instrument.tracing.events.CallEvent;
-import edu.berkeley.cs.jqf.instrument.tracing.events.ReturnEvent;
 import edu.berkeley.cs.jqf.instrument.tracing.events.TraceEvent;
-import edu.berkeley.cs.jqf.instrument.tracing.events.TraceEventVisitor;
 
 import static java.lang.Math.ceil;
 import static java.lang.Math.log;
@@ -78,148 +71,128 @@ import static java.lang.Math.log;
  *
  * @author Rohan Padhye
  */
-public class ZestGuidance implements Guidance, TraceEventVisitor {
+public class ZestGuidance implements Guidance {
 
     // Currently, we only support single-threaded applications
     // This field is used to ensure that
-    private Thread appThread;
-
-    /** The last event handled by this guidance */
-    private TraceEvent lastEvent;
-
-    /** The execution indexing logic. */
-    private ExecutionIndexingState eiState;
+    protected Thread appThread;
 
     /** A pseudo-random number generator for generating fresh values. */
-    private Random random = new Random();
+    protected Random random = new Random();
 
     /** The name of the test for display purposes. */
-    private final String testName;
+    protected final String testName;
 
     // ------------ ALGORITHM BOOKKEEPING ------------
 
     /** The max amount of time to run for, in milli-seconds */
-    private final long maxDurationMillis;
+    protected final long maxDurationMillis;
 
     /** The number of trials completed. */
-    private long numTrials = 0;
+    protected long numTrials = 0;
 
     /** The number of valid inputs. */
-    private long numValid = 0;
+    protected long numValid = 0;
 
     /** The directory where fuzzing results are written. */
-    private final File outputDirectory;
+    protected final File outputDirectory;
 
     /** The directory where saved inputs are written. */
-    private File savedInputsDirectory;
+    protected File savedInputsDirectory;
 
     /** The directory where saved inputs are written. */
-    private File savedFailuresDirectory;
+    protected File savedFailuresDirectory;
 
     /** Set of saved inputs to fuzz. */
-    private ArrayList<Input> savedInputs = new ArrayList<>();
+    protected ArrayList<Input> savedInputs = new ArrayList<>();
 
     /** Queue of seeds to fuzz. */
-    private Deque<SeedInput> seedInputs = new ArrayDeque<>();
+    protected Deque<SeedInput> seedInputs = new ArrayDeque<>();
 
     /** Current input that's running -- valid after getInput() and before handleResult(). */
-    private Input<?> currentInput;
+    protected Input<?> currentInput;
 
     /** Index of currentInput in the savedInputs -- valid after seeds are processed (OK if this is inaccurate). */
-    private int currentParentInputIdx = 0;
+    protected int currentParentInputIdx = 0;
 
     /** Number of mutated inputs generated from currentInput. */
-    private int numChildrenGeneratedForCurrentParentInput = 0;
+    protected int numChildrenGeneratedForCurrentParentInput = 0;
 
     /** Number of cycles completed (i.e. how many times we've reset currentParentInputIdx to 0. */
-    private int cyclesCompleted = 0;
+    protected int cyclesCompleted = 0;
 
     /** Number of favored inputs in the last cycle. */
-    private int numFavoredLastCycle = 0;
+    protected int numFavoredLastCycle = 0;
 
     /** Blind fuzzing -- if true then the queue is always empty. */
-    private boolean blind;
+    protected boolean blind;
 
     /** Number of saved inputs.
      *
      * This is usually the same as savedInputs.size(),
      * but we do not really save inputs in TOTALLY_RANDOM mode.
      */
-    private int numSavedInputs = 0;
+    protected int numSavedInputs = 0;
 
     /** Coverage statistics for a single run. */
-    private Coverage runCoverage = new Coverage();
+    protected Coverage runCoverage = new Coverage();
 
     /** Cumulative coverage statistics. */
-    private Coverage totalCoverage = new Coverage();
+    protected Coverage totalCoverage = new Coverage();
 
     /** Cumulative coverage for valid inputs. */
-    private Coverage validCoverage = new Coverage();
+    protected Coverage validCoverage = new Coverage();
 
     /** The maximum number of keys covered by any single input found so far. */
-    private int maxCoverage = 0;
+    protected int maxCoverage = 0;
 
     /** A mapping of coverage keys to inputs that are responsible for them. */
-    private Map<Object, Input> responsibleInputs = new HashMap<>(totalCoverage.size());
+    protected Map<Object, Input> responsibleInputs = new HashMap<>(totalCoverage.size());
 
     /** The set of unique failures found so far. */
-    Set<List<StackTraceElement>> uniqueFailures = new HashSet<>();
-
-    /**
-     * A map of execution contexts (call stacks) to locations in saved inputs with those contexts.
-     *
-     * This is a nifty data structure for quickly finding candidates for input splicing.
-     */
-    private Map<ExecutionContext, ArrayList<InputLocation>> ecToInputLoc
-            = new ProducerHashMap<>(() -> new ArrayList<>());
+    protected Set<List<StackTraceElement>> uniqueFailures = new HashSet<>();
 
     // ---------- LOGGING / STATS OUTPUT ------------
 
     /** Whether to print log statements to stderr (debug option; manually edit). */
-    private final boolean verbose = true;
+    protected final boolean verbose = true;
 
     /** A system console, which is non-null only if STDOUT is a console. */
-    private final Console console = System.console();
+    protected final Console console = System.console();
 
     /** Time since this guidance instance was created. */
-    private final Date startTime = new Date();
+    protected final Date startTime = new Date();
 
     /** Time at last stats refresh. */
-    private Date lastRefreshTime = startTime;
+    protected Date lastRefreshTime = startTime;
 
     /** Total execs at last stats refresh. */
-    private long lastNumTrials = 0;
+    protected long lastNumTrials = 0;
 
     /** Minimum amount of time (in millis) between two stats refreshes. */
-    private static final long STATS_REFRESH_TIME_PERIOD = 300;
+    protected static final long STATS_REFRESH_TIME_PERIOD = 300;
 
     /** The file where log data is written. */
-    private File logFile;
+    protected File logFile;
 
     /** The file where saved plot data is written. */
-    private File statsFile;
+    protected File statsFile;
 
     /** The currently executing input (for debugging purposes). */
-    private File currentInputFile;
-
-    /** Whether to print the fuzz config to the stats screen. */
-    private static boolean SHOW_CONFIG = false;
+    protected File currentInputFile;
 
     // ------------- TIMEOUT HANDLING ------------
 
     /** Timeout for an individual run. */
-    private long singleRunTimeoutMillis;
+    protected long singleRunTimeoutMillis;
 
     /** Date when last run was started. */
-    private Date runStart;
+    protected Date runStart;
 
     /** Number of conditional jumps since last run was started. */
-    private long branchCount;
+    protected long branchCount;
 
     // ------------- FUZZING HEURISTICS ------------
-
-    /** Whether to use real execution indexes as opposed to flat numbering. */
-    static final boolean DISABLE_EXECUTION_INDEXING = !Boolean.getBoolean("jqf.ei.ENABLE_EXECUTION_INDEXING");
 
     /** Whether to save only valid inputs **/
     static final boolean SAVE_ONLY_VALID = Boolean.getBoolean("jqf.ei.SAVE_ONLY_VALID");
@@ -242,25 +215,16 @@ public class ZestGuidance implements Guidance, TraceEventVisitor {
     /** Mean number of contiguous bytes to mutate in each mutation. */
     static final double MEAN_MUTATION_SIZE = 4.0; // Bytes
 
-    /** Max number of contiguous bytes to splice in from another input during the splicing stage. */
-    static final int MAX_SPLICE_SIZE = 64; // Bytes
-
-    /** Whether to splice only in the same sub-tree */
-    static final boolean SPLICE_SUBTREE = Boolean.getBoolean("jqf.ei.SPLICE_SUBTREE");
-
     /** Whether to save inputs that only add new coverage bits (but no new responsibilities). */
     static final boolean SAVE_NEW_COUNTS = true;
 
     /** Whether to steal responsibility from old inputs (this increases computation cost). */
     static final boolean STEAL_RESPONSIBILITY = Boolean.getBoolean("jqf.ei.STEAL_RESPONSIBILITY");
 
-    /** Probability of splicing in getOrGenerateFresh() */
-    static final double DEMAND_DRIVEN_SPLICING_PROBABILITY = 0;
-
     /**
-     * @param testName the name of test to display on the status screen
-     * Creates a new execution-index-parametric guidance.
+     * Creates a new guidance instance.
      *
+     * @param testName the name of test to display on the status screen
      * @param duration the amount of time to run fuzzing for, where
      *                 {@code null} indicates unlimited time.
      * @param outputDirectory the directory where fuzzing results will be written
@@ -286,6 +250,8 @@ public class ZestGuidance implements Guidance, TraceEventVisitor {
     }
 
     /**
+     * Creates a new guidance instance.
+     *
      * @param testName the name of test to display on the status screen
      * @param duration the amount of time to run fuzzing for, where
      *                 {@code null} indicates unlimited time.
@@ -416,27 +382,11 @@ public class ZestGuidance implements Guidance, TraceEventVisitor {
 
         console.printf("\033[2J");
         console.printf("\033[H");
-        if (blind) {
-            console.printf("Generator-based random fuzzing (no guidance)\n");
-            console.printf("--------------------------------------------\n");
-        } else {
-            console.printf("Semantic Fuzzing with Zest\n");
-            console.printf("--------------------------\n");
-        }
+        console.printf(this.getTitle() + "\n");
         if (this.testName != null) {
             console.printf("Test name:            %s\n", this.testName);
         }
         console.printf("Results directory:    %s\n", this.outputDirectory.getAbsolutePath());
-        if (SHOW_CONFIG) {
-            if (blind) {
-                console.printf("Config:               TOTALLY_RANDOM\n");
-            } else {
-                console.printf("Config:               DISABLE_EXECUTION_INDEXING = %s,\n" +
-                                "                      STEAL_RESPONSIBILITY       = %s,\n" +
-                                "                      SPLICE_SUBTREE             = %s\n\n",
-                        DISABLE_EXECUTION_INDEXING, STEAL_RESPONSIBILITY, SPLICE_SUBTREE);
-            }
-        }
         console.printf("Elapsed time:         %s (%s)\n", millisToDuration(elapsedMilliseconds),
                 maxDurationMillis == Long.MAX_VALUE ? "no time limit" : ("max " + millisToDuration(maxDurationMillis)));
         console.printf("Number of executions: %,d\n", numTrials);
@@ -455,6 +405,17 @@ public class ZestGuidance implements Guidance, TraceEventVisitor {
                 numValid, numTrials-numValid, nonZeroValidFraction);
         appendLineToFile(statsFile, plotData);
 
+    }
+
+    /** Returns the banner to be displayed on the status screen */
+    protected String getTitle() {
+        if (blind) {
+            return  "Generator-based random fuzzing (no guidance)\n" +
+                    "--------------------------------------------\n";
+        } else {
+            return  "Semantic Fuzzing with Zest\n" +
+                    "--------------------------\n";
+        }
     }
 
     public void setBlind(boolean blind) {
@@ -478,7 +439,8 @@ public class ZestGuidance implements Guidance, TraceEventVisitor {
         return target;
     }
 
-    private void completeCycle() {
+    /** Handles the end of fuzzing cycle (i.e., having gone through the entire queue) */
+    protected void completeCycle() {
         // Increment cycle count
         cyclesCompleted++;
         infoLog("\n# Cycle " + cyclesCompleted + " completed.");
@@ -501,25 +463,44 @@ public class ZestGuidance implements Guidance, TraceEventVisitor {
             throw new AssertionError("Responsibilty mistmatch");
         }
 
-        // Refresh ecToInputLoc so that subsequent splices are only from favored inputs
-        ecToInputLoc.clear();
-        for (Input input : savedInputs) {
-            if (input.isFavored()) {
-                mapEcToInputLoc(input);
-            }
-        }
-
         // Break log after cycle
         infoLog("\n\n\n");
+    }
+
+    /** Spawns a new input from thin air (i.e., actually random) */
+    protected Input<?> createFeshInput() {
+        return new LinearInput();
+    }
+
+    /**
+     * Returns an InputStream that delivers parameters to the generators.
+     *
+     * Note: The variable `currentInput` has been set to point to the input
+     * to mutate.
+     */
+    protected InputStream createParameterStream() {
+        // Return an input stream that reads bytes from a linear array
+        return new InputStream() {
+            int bytesRead = 0;
+
+            @Override
+            public int read() throws IOException {
+                assert currentInput instanceof LinearInput : "ZestGuidance should only mutate LinearInput(s)";
+
+                // For linear inputs, get with key = bytesRead (which is then incremented)
+                LinearInput linearInput = (LinearInput) currentInput;
+                // Attempt to get a value from the list, or else generate a random value
+                int ret = linearInput.getOrGenerateFresh(bytesRead++, random);
+                // infoLog("read(%d) = %d", bytesRead, ret);
+                return ret;
+            }
+        };
     }
 
     @Override
     public InputStream getInput() throws GuidanceException {
         // Clear coverage stats for this run
         runCoverage.clear();
-
-        // Reset execution index state
-        eiState = new ExecutionIndexingState();
 
         // Choose an input to execute based on state of queues
         if (!seedInputs.isEmpty()) {
@@ -537,7 +518,7 @@ public class ZestGuidance implements Guidance, TraceEventVisitor {
 
             // Make fresh input using either list or maps
             infoLog("Spawning new input from thin air");
-            currentInput = DISABLE_EXECUTION_INDEXING ? new LinearInput() : new MappedInput();
+            currentInput = createFeshInput();
         } else {
             // The number of children to produce is determined by how much of the coverage
             // pool this parent input hits
@@ -572,45 +553,7 @@ public class ZestGuidance implements Guidance, TraceEventVisitor {
         }
 
 
-        // Return an input stream that uses the EI map
-        return new InputStream() {
-            int bytesRead = 0;
-
-            @Override
-            public int read() throws IOException {
-
-                // lastEvent must not be null
-                if (DISABLE_EXECUTION_INDEXING == false && lastEvent == null) {
-                    throw new IOException("Could not compute execution index; no instrumentation?");
-                }
-
-                // For linear inputs, get with key = bytesRead (which is then incremented)
-                if (currentInput instanceof LinearInput) {
-                    LinearInput linearInput = (LinearInput) currentInput;
-                    // Attempt to get a value from the list, or else generate a random value
-                    int ret = linearInput.getOrGenerateFresh(bytesRead++, random);
-                    // infoLog("read(%d) = %d", bytesRead, ret);
-                    return ret;
-                }
-
-                // For mapped inputs, make a suitable execution index
-                else {
-                    MappedInput mappedInput = (MappedInput) currentInput;
-
-                    // Get the execution index of the last event
-                    ExecutionIndex executionIndex = eiState.getExecutionIndex(lastEvent);
-
-                    // Attempt to get a value from the map, or else generate a random value
-                    int value = mappedInput.getOrGenerateFresh(executionIndex, random);
-
-                    // Keep track of how many bytes were read in this input
-                    bytesRead++;
-
-                    return value;
-
-                }
-            }
-        };
+        return createParameterStream();
     }
 
     @Override
@@ -817,7 +760,8 @@ public class ZestGuidance implements Guidance, TraceEventVisitor {
 
     }
 
-    private void saveCurrentInput(Set<Object> responsibilities, String why) throws IOException {
+    /* Saves an interesting input to the queue. */
+    protected void saveCurrentInput(Set<Object> responsibilities, String why) throws IOException {
 
         // First, save to disk (note: we issue IDs to everyone, but only write to disk  if valid)
         int newInputIdx = numSavedInputs++;
@@ -861,21 +805,6 @@ public class ZestGuidance implements Guidance, TraceEventVisitor {
             responsibleInputs.put(b, currentInput);
         }
 
-        // Fifth, map executions to input locations for splicing
-        mapEcToInputLoc(currentInput);
-
-    }
-
-    private void mapEcToInputLoc(Input input) {
-        if (input instanceof MappedInput) {
-            MappedInput mappedInput = (MappedInput) input;
-            for (int offset = 0; offset < mappedInput.size(); offset++) {
-                ExecutionIndex ei = mappedInput.orderedKeys.get(offset);
-                ExecutionContext ec = new ExecutionContext(ei);
-                ecToInputLoc.get(ec).add(new InputLocation(mappedInput, offset));
-            }
-        }
-
     }
 
 
@@ -890,15 +819,8 @@ public class ZestGuidance implements Guidance, TraceEventVisitor {
         return this::handleEvent;
     }
 
-    private void handleEvent(TraceEvent e) {
-        // Set last event to this event
-        lastEvent = e;
-
-        // Update execution indexing logic
-        if (!DISABLE_EXECUTION_INDEXING) {
-            e.applyVisitor(this);
-        }
-
+    /** Handles a trace event generated during test execution */
+    protected void handleEvent(TraceEvent e) {
         // Collect totalCoverage
         runCoverage.handleEvent(e);
         // Check for possible timeouts every so often
@@ -911,15 +833,6 @@ public class ZestGuidance implements Guidance, TraceEventVisitor {
         }
     }
 
-    @Override
-    public void visitCallEvent(CallEvent c) {
-        eiState.pushCall(c);
-    }
-
-    @Override
-    public void visitReturnEvent(ReturnEvent r) {
-        eiState.popReturn(r);
-    }
 
     /**
      * Returns a reference to the coverage statistics.
@@ -1032,7 +945,7 @@ public class ZestGuidance implements Guidance, TraceEventVisitor {
          *
          * @return
          */
-        private boolean isFavored() {
+        public boolean isFavored() {
             return responsibilities.size() > 0;
         }
 
@@ -1046,7 +959,7 @@ public class ZestGuidance implements Guidance, TraceEventVisitor {
          * @param mean the mean of the distribution
          * @return a randomly sampled value
          */
-        protected static int sampleGeometric(Random random, double mean) {
+        public static int sampleGeometric(Random random, double mean) {
             double p = 1 / mean;
             double uniform = random.nextDouble();
             return (int) ceil(log(1 - uniform) / log(1 - p));
@@ -1166,458 +1079,6 @@ public class ZestGuidance implements Guidance, TraceEventVisitor {
         }
     }
 
-
-    /**
-     * A candidate test input represented as a map from execution indices
-     * to integer values.
-     *
-     * <p>When a quickcheck-like generator requests a new ``random'' byte,
-     * the current execution index is used to retrieve the input from
-     * this input map (a fresh value is generated and stored in the map
-     * if the key is not mapped).</p>
-     *
-     * <p>Inputs should not be publicly mutable. The only way to mutate
-     * an input is via the {@link #fuzz} method which produces a new input
-     * object with some values mutated.</p>
-     */
-    public class MappedInput extends Input<ExecutionIndex> {
-
-        /**
-         * Whether this input has been executed.
-         *
-         * When this field is {@code false}, the field {@link #orderedKeys}
-         * is not yet populated and must not be used. When this field is {@code true},
-         * the input should be considered immutable and neither {@link #orderedKeys} nor
-         * {@link #valuesMap} must be modified.
-         */
-        protected boolean executed = false;
-
-        /** A map from execution indexes to the byte (0-255) to be returned at that index. */
-        protected LinkedHashMap<ExecutionIndex, Integer> valuesMap;
-
-        /**
-         * A list of execution indexes that are actually requested by the test program when
-         * executed with this input.
-         *
-         * <p>This list is initially empty, and is populated at the end of the run, after which
-         * it is frozen. The list of keys are in order of their occurrence in the execution
-         * trace and can therefore be used to serialize the map into a sequence of bytes.</p>
-         *
-         */
-        protected ArrayList<ExecutionIndex> orderedKeys = new ArrayList<>();
-
-
-        private List<InputPrefixMapping> demandDrivenSpliceMap = new ArrayList<>();
-
-        /**
-         * Create an empty input map.
-         */
-        public MappedInput() {
-            super();
-            valuesMap = new LinkedHashMap<>();
-        }
-
-        /**
-         * Create a copy of an existing input map.
-         *
-         * @param toClone the input map to clone
-         */
-        public MappedInput(MappedInput toClone) {
-            super(toClone);
-            valuesMap = new LinkedHashMap<>(toClone.valuesMap);
-        }
-
-        /**
-         * Returns the size of this input, in terms of number of bytes
-         * in its value map.
-         *
-         * @return the size of this input
-         */
-        public final int size() {
-            return valuesMap.size();
-        }
-
-        /**
-         * Returns the byte mapped by this input at a given offset.
-         *
-         * @param offset the byte offset in the input
-         * @return the byte value at that offset
-         *
-         * @throws IndexOutOfBoundsException if the offset is negative or
-         *      larger than {@link #size}()-1
-         * @throws IllegalStateException if this method is called before the input
-         *                               has been executed
-         */
-        private final int getValueAtOffset(int offset) throws IndexOutOfBoundsException, IllegalStateException {
-            if (!executed) {
-                throw new IllegalStateException("Cannot get with offset before execution");
-            }
-
-            // Return the mapping for the execution index queried at the offset
-            ExecutionIndex ei = orderedKeys.get(offset);
-            return valuesMap.get(ei);
-        }
-
-
-        /**
-         * Returns the execution index mapped by this input at a given offset.
-         *
-         * @param offset the byte offset in the input
-         * @return the execution index value at that offset
-         *
-         * @throws IndexOutOfBoundsException if the offset is negative or
-         *      larger than {@link #size}()-1
-         * @throws IllegalStateException if this method is called before the input
-         *                               has been executed
-         */
-        private final ExecutionIndex getKeyAtOffset(int offset) throws IndexOutOfBoundsException, IllegalStateException {
-            if (!executed) {
-                throw new IllegalStateException("Cannot get with offset before execution");
-            }
-
-            // Return the execution index queried at the offset
-            return orderedKeys.get(offset);
-        }
-
-        private InputPrefixMapping getInputPrefixMapping(ExecutionIndex ei) {
-            for (InputPrefixMapping ipm : demandDrivenSpliceMap) {
-                if (ei.hasPrefix(ipm.targetPrefix)) {
-                    return ipm;
-                }
-            }
-            return null;
-        }
-
-
-        /**
-         * Retrieve a value for an execution index if mapped, else generate
-         * a fresh value.
-         *
-         * @param key    the execution index of the trace event requesting a new byte
-         * @param random the PRNG
-         * @return the value to return to the quickcheck-like generator
-         * @throws IllegalStateException if this method is called after the input
-         *                               has been executed
-         */
-        @Override
-        public int getOrGenerateFresh(ExecutionIndex key, Random random) throws IllegalStateException {
-            if (executed) {
-                throw new IllegalStateException("Cannot generate fresh values after execution");
-            }
-
-            // If we reached a limit, then just return EOF
-            if (orderedKeys.size() >= MAX_INPUT_SIZE) {
-                return -1;
-            }
-
-            // Try to get existing values
-            Integer val = valuesMap.get(key);
-
-            // If not, generate a new value
-            if (val == null) {
-                InputPrefixMapping ipm;
-
-                // If we have an input prefix mapping for this execution index,
-                // then splice from the source input
-                if ((ipm = getInputPrefixMapping(key)) != null) {
-                    Prefix sourcePrefix = ipm.sourcePrefix;
-                    Suffix sourceSuffix = ipm.sourcePrefix.getEi().getSuffixOfPrefix(sourcePrefix);
-                    ExecutionIndex sourceEi = new ExecutionIndex(sourcePrefix, sourceSuffix);
-                    // The value can be taken from the source
-                    val = ipm.sourceInput.getValueAtKey(sourceEi);
-                }
-
-                // If we could not splice or were unsuccessful, try to generate a new input
-                if (val == null) {
-                    if (GENERATE_EOF_WHEN_OUT) {
-                        return -1;
-                    }
-                    if (random.nextDouble() < DEMAND_DRIVEN_SPLICING_PROBABILITY) {
-                        // TODO: Find a random inputLocation with same EC,
-                        // extract common suffix of sourceEi and targetEi,
-                        // and map targetPrefix to sourcePrefix in the IPM
-
-
-                    } else {
-                        // Just generate a random input
-                        val = random.nextInt(256);
-                    }
-                }
-
-                // Put the new value into the map
-                assert (val != null);
-
-                valuesMap.put(key, val);
-            }
-
-            // Mark this key as visited
-            orderedKeys.add(key);
-
-            return val;
-        }
-
-
-        /**
-         * Gets the byte mapped by this input at a given execution index.
-         *
-         * @param ei the execution index
-         * @return the value mapped for this index, or {@code null} if no such mapping exists
-         *
-         * @throws IndexOutOfBoundsException if the offset is negative or
-         *      larger than {@link #size}()-1
-         */
-        protected final Integer getValueAtKey(ExecutionIndex ei) throws IndexOutOfBoundsException {
-            return valuesMap.get(ei);
-        }
-
-        /**
-         * Sets the byte mapped by this input at a given execution index.
-         *
-         * @param ei  the execution index at which to insert
-         * @param val the byte to insert
-         *
-         * @throws IndexOutOfBoundsException if the offset is negative or
-         *      larger than {@link #size}()-1
-         * @throws IllegalStateException if this method is called after the input
-         *                               has been executed
-         */
-        protected final void setValueAtKey(ExecutionIndex ei, int val) throws IndexOutOfBoundsException, IllegalStateException {
-            if (executed) {
-                throw new IllegalStateException("Cannot set value before execution");
-            }
-
-            valuesMap.put(ei, val);
-        }
-
-        /**
-         * Trims the input map of all keys that were never actually requested since
-         * its construction.
-         *
-         * <p>Although this operation mutates the underlying object, the effect should
-         * not be externally visible (at least as long as the test executions are
-         * deterministic).</p>
-         */
-        @Override
-        public void gc() {
-            LinkedHashMap<ExecutionIndex, Integer> newMap = new LinkedHashMap<>();
-            for (ExecutionIndex key : orderedKeys) {
-                newMap.put(key, valuesMap.get(key));
-            }
-            valuesMap = newMap;
-
-            // Set the `executed` flag
-            executed = true;
-        }
-
-        /**
-         * Return a new input derived from this one with some values
-         * mutated.
-         *
-         * Pass-through to {@link #fuzz(Random, Map)}
-         *
-         */
-        @Override
-        public Input fuzz(Random random) {
-            return fuzz(random, ZestGuidance.this.ecToInputLoc);
-        }
-
-        /**
-         * Return a new input derived from this one with some values
-         * mutated.
-         *
-         * <p>This method performs one or both of random mutations
-         * and splicing.</p>
-         *
-         * <p>Random mutations are done by performing M
-         * mutation operations each on a random contiguous sequence of N bytes,
-         * where M and N are sampled from a geometric distribution with mean
-         * {@link #MEAN_MUTATION_COUNT} and {@link #MEAN_MUTATION_SIZE}
-         * respectively.</p>
-         *
-         * <p>Splicing is performed by first randomly choosing a location and
-         * its corresponding execution context in this input's value map, and then
-         * copying a contiguous sequence of up to Z bytes from another input,
-         * starting with a location that also maps the same execution context.
-         * Here, Z is sampled from a uniform distribution from 0 to
-         * {@link #MAX_SPLICE_SIZE}.</p>
-         *
-         * @param random the PRNG
-         * @return a newly fuzzed input
-         */
-        protected MappedInput fuzz(Random random, Map<ExecutionContext, ArrayList<InputLocation>> ecToInputLoc) {
-            // Derive new input from this object as source
-            MappedInput newInput = new MappedInput(this);
-
-            // Maybe try splicing
-            boolean splicingDone = false;
-
-            // Only splice if we have been provided the ecToInputLoc
-            if (ecToInputLoc != null) {
-
-                // TODO: Do we really want splicing to be this frequent?
-                if (random.nextBoolean()) {
-                    final int MIN_TARGET_ATTEMPTS = 3;
-                    final int MAX_TARGET_ATTEMPTS = 6;
-
-                    int targetAttempts = MIN_TARGET_ATTEMPTS;
-
-                    outer: for (int targetAttempt = 1; targetAttempt < targetAttempts; targetAttempt++) {
-
-                        // Choose an execution context at which to splice at
-                        // Note: We get EI and value from `this` rather than `newInput`
-                        // because `this` has already been executed
-                        int targetOffset = random.nextInt(newInput.valuesMap.size());
-                        ExecutionIndex targetEi = this.getKeyAtOffset(targetOffset);
-
-                        ExecutionContext targetEc = new ExecutionContext(targetEi);
-                        int valueAtTarget = this.getValueAtOffset(targetOffset);
-
-                        // Find a suitable input location to splice from
-                        ArrayList<InputLocation> inputLocations = ecToInputLoc.get(targetEc);
-
-                        // If this was a bad choice of target, try again without penalty if possible
-                        if (inputLocations.size() == 0) {
-                            // Try to increase the loop bound a little bit to get another chance
-                            targetAttempts = Math.min(targetAttempts+1, MAX_TARGET_ATTEMPTS);
-                            continue;
-                        }
-
-                        InputLocation inputLocation;
-
-                        // Try a bunch of times
-                        for (int attempt = 1; attempt <= 10; attempt++) {
-
-                            // Get a candidate source location with the same execution context
-                            inputLocation = inputLocations.get(random.nextInt(inputLocations.size()));
-                            MappedInput sourceInput = inputLocation.input;
-                            int sourceOffset = inputLocation.offset;
-
-
-                            // Do not splice with ourselves
-                            if (sourceInput == this) {
-                                continue;
-                            }
-
-                            // Do not splice if the first value is the same in source and target
-                            if (sourceInput.getValueAtOffset(sourceOffset) == valueAtTarget) {
-                                continue;
-                            }
-
-                            int splicedBytes = 0;
-                            if (!DISABLE_EXECUTION_INDEXING && SPLICE_SUBTREE) {
-                                // Do not splice if there is no common suffix between EI of source and target
-                                ExecutionIndex sourceEi = sourceInput.getKeyAtOffset(sourceOffset);
-                                Suffix suffix = targetEi.getCommonSuffix(sourceEi);
-                                if (suffix.size() == 0) {
-                                    continue;
-                                }
-
-                                // Extract the source and target prefixes
-                                Prefix sourcePrefix = sourceEi.getPrefixOfSuffix(suffix);
-                                Prefix targetPrefix = targetEi.getPrefixOfSuffix(suffix);
-                                assert (sourcePrefix.size() == targetPrefix.size());
-
-                                // OK, this looks good. Let's splice!
-                                int srcIdx = sourceOffset;
-                                while (srcIdx < sourceInput.size()) {
-                                    ExecutionIndex candidateEi = sourceInput.getKeyAtOffset(srcIdx);
-                                    if (candidateEi.hasPrefix(sourcePrefix) == false) {
-                                        // We are no more in the same sub-tree as sourceEi
-                                        break;
-                                    }
-                                    Suffix spliceSuffix = candidateEi.getSuffixOfPrefix(sourcePrefix);
-                                    ExecutionIndex spliceEi = new ExecutionIndex(targetPrefix, spliceSuffix);
-                                    newInput.valuesMap.put(spliceEi, sourceInput.valuesMap.get(candidateEi));
-
-                                    srcIdx++;
-                                }
-                                splicedBytes = srcIdx - sourceOffset;
-                            } else {
-
-                                int spliceSize = 1 + random.nextInt(MAX_SPLICE_SIZE);
-                                int src = sourceOffset;
-                                int tgt = targetOffset;
-                                int srcSize = sourceInput.size();
-                                int tgtSize = newInput.size();
-                                while (splicedBytes < spliceSize && src < srcSize && tgt < tgtSize) {
-                                    int val = sourceInput.getValueAtOffset(src);
-                                    ExecutionIndex key = this.getKeyAtOffset(tgt);
-                                    newInput.setValueAtKey(key, val);
-
-                                    splicedBytes++;
-                                    src++;
-                                    tgt++;
-                                }
-                            }
-
-                            // Complete splicing
-                            splicingDone = true;
-                            newInput.desc += String.format(",splice:%06d:%d@%d->%d", sourceInput.id, splicedBytes,
-                                    sourceOffset, targetOffset);
-
-                            break outer; // Stop more splicing attempts!
-
-                        }
-                    }
-                }
-            }
-
-            // Maybe do random mutations
-            if (splicingDone == false || random.nextBoolean()) {
-
-                // Stack a bunch of mutations
-                int numMutations = sampleGeometric(random, MEAN_MUTATION_COUNT);
-                newInput.desc += ",havoc:"+numMutations;
-
-                boolean setToZero = random.nextDouble() < 0.1; // one out of 10 times
-
-                for (int mutation = 1; mutation <= numMutations; mutation++) {
-
-                    // Select a random offset and size
-                    int offset = random.nextInt(newInput.valuesMap.size());
-                    int mutationSize = sampleGeometric(random, MEAN_MUTATION_SIZE);
-
-                    // desc += String.format(":%d@%d", mutationSize, idx);
-
-                    // Iterate over all entries in the value map
-                    Iterator<Map.Entry<ExecutionIndex, Integer>> entryIterator
-                            = newInput.valuesMap.entrySet().iterator();
-                    for (int i = 0; entryIterator.hasNext(); i++) {
-                        Map.Entry<ExecutionIndex, Integer> e = entryIterator.next();
-                        // Only mutate `mutationSize` contiguous entries from
-                        // the randomly selected `idx`.
-                        if (i >= offset && i < (offset + mutationSize)) {
-                            // Apply a random mutation
-                            int mutatedValue = setToZero ? 0 : random.nextInt(256);
-                            e.setValue(mutatedValue);
-                        }
-                    }
-                }
-            }
-
-            return newInput;
-
-        }
-
-        @Override
-        public Iterator<Integer> iterator() {
-            return new Iterator<Integer>() {
-
-                Iterator<ExecutionIndex> keyIt = orderedKeys.iterator();
-
-                @Override
-                public boolean hasNext() {
-                    return keyIt.hasNext();
-                }
-
-                @Override
-                public Integer next() {
-                    return valuesMap.get(keyIt.next());
-                }
-            };
-        }
-    }
-
     public class SeedInput extends LinearInput {
         final File seedFile;
         final InputStream in;
@@ -1666,27 +1127,5 @@ public class ZestGuidance implements Guidance, TraceEventVisitor {
 
     }
 
-
-    static class InputLocation {
-        private final MappedInput input;
-        private final int offset;
-
-        InputLocation(MappedInput input, int offset) {
-            this.input = input;
-            this.offset = offset;
-        }
-    }
-
-    static class InputPrefixMapping {
-        private final MappedInput sourceInput;
-        private final Prefix sourcePrefix;
-        private final Prefix targetPrefix;
-
-        InputPrefixMapping(MappedInput sourceInput, Prefix sourcePrefix, Prefix targetPrefix) {
-            this.sourceInput = sourceInput;
-            this.sourcePrefix = sourcePrefix;
-            this.targetPrefix = targetPrefix;
-        }
-    }
 
 }
