+++ scripts/janala.conf
@@ -1,2 +1,2 @@
 janala.excludes=java/,com/sun/proxy/,com/intellij/,edu/berkeley/cs/jqf/,org/junit/,com/pholser/junit/quickcheck/,ru/vyarus/java/generics/resolver/,org/javaruntype/,ognl,org/hamcrest/,org/omg/,org/netbeans/,org/mozilla/javascript/gen
-janala.includes=edu/berkeley/cs/jqf/examples,java/text,java/time,com/sun/imageio
+janala.includes=edu/berkeley/cs/jqf/examples,java/text,java/time,com/sun/imageio,com/pholser/junit/quickcheck/internal,com/pholser/junit/quickcheck/generator
